aufs: support linux-2.6.30, temporary workaround for relatime option
The mount option 'relatime' becomes set by default.
Although it is VFS option, mount(8) may pass it to aufs (or all
filesystems) since old version of mount(8) may not know about
relatime. When common VFS option is passed to aufs, aufs does NOT handle
it and will return an error as "unknown option".
It may be unconvenient for users. So here is a temporary workaround to
ignore relatime silently.

Signed-off-by: J. R. Okajima <hooanon05@yahoo.co.jp>