aufs: new lockclass for copyup whiteout
When a process writes a file which exists in the lower readonly branch
and was already removed with its parent dir, aufs copies it up by
au_sio_cpup_wh().
Since the parent dir is already removed, the function has to copy the
file to the internal temporary dir. At this point, new lock class is
necessary.

Signed-off-by: J. R. Okajima <hooanon05@yahoo.co.jp>